Refactored admin panel into blueprints
