+++ app/templates/base.html
@@ -24,7 +24,7 @@
                 </li>
                 {% if current_user.role.role == 'admin' %}
                     <li class="nav-item">
-                        <a class="nav-link" href="{{ url_for('admin') }}">Admin</a>
+                        <a class="nav-link" href="{{ url_for('admin.admin') }}">Admin</a>
                     </li>
                 {% endif %}
                 <li class="nav-item">
